+++ app/helpers/grouping_helper.rb
@@ -18,12 +18,4 @@ module GroupingHelper
                 :'data-y' => y,
                 style: "transform: translate(#{x}px, #{y}px)")
   end
-
-  class ActionView::Helpers::FormBuilder
-    include ActionView::Helpers::FormTagHelper
-
-    def grouping_select(method, **html_options)
-      collection_select method, Grouping.all, :id, :name, { include_blank: '' }, html_options
-    end
-  end
 end
